+++ src/Symfony/Component/Serializer/Normalizer/AbstractObjectNormalizer.php
@@ -139,8 +139,6 @@ abstract class AbstractObjectNormalizer extends AbstractNormalizer
     }
 
     /**
-     * @param array $context
-     *
      * @return bool
      */
     public function supportsNormalization(mixed $data, string $format = null /* , array $context = [] */)
@@ -290,14 +288,10 @@ abstract class AbstractObjectNormalizer extends AbstractNormalizer
 
     /**
      * Gets the attribute value.
-     *
-     * @return mixed
      */
     abstract protected function getAttributeValue(object $object, string $attribute, string $format = null, array $context = []);
 
     /**
-     * @param array $context
-     *
      * @return bool
      */
     public function supportsDenormalization(mixed $data, string $type, string $format = null /* , array $context = [] */)
@@ -305,9 +299,6 @@ abstract class AbstractObjectNormalizer extends AbstractNormalizer
         return class_exists($type) || (interface_exists($type, false) && null !== $this->classDiscriminatorResolver?->getMappingForClass($type));
     }
 
-    /**
-     * @return mixed
-     */
     public function denormalize(mixed $data, string $type, string $format = null, array $context = [])
     {
         if (!isset($context['cache_key'])) {
@@ -439,11 +430,9 @@ abstract class AbstractObjectNormalizer extends AbstractNormalizer
                 return null;
             }
 
-            $collectionValueType = $type->isCollection() ? $type->getCollectionValueTypes()[0] ?? null : null;
-
             // Fix a collection that contains the only one element
             // This is special to xml format only
-            if ('xml' === $format && null !== $collectionValueType && (!\is_array($data) || !\is_int(key($data)))) {
+            if ('xml' === $format && \count($type->getCollectionValueTypes()) > 0 && (!\is_array($data) || !\is_int(key($data)))) {
                 $data = [$data];
             }
 
@@ -499,41 +488,33 @@ abstract class AbstractObjectNormalizer extends AbstractNormalizer
                     }
                 }
 
-                if (null !== $collectionValueType && Type::BUILTIN_TYPE_OBJECT === $collectionValueType->getBuiltinType()) {
-                    $builtinType = Type::BUILTIN_TYPE_OBJECT;
-                    $class = $collectionValueType->getClassName().'[]';
+                $builtinType = $type->getBuiltinType();
+                $class = $type->getClassName();
 
-                    if (\count($collectionKeyType = $type->getCollectionKeyTypes()) > 0) {
-                        $context['key_type'] = \count($collectionKeyType) > 1 ? $collectionKeyType : $collectionKeyType[0];
-                    }
-
-                    $context['value_type'] = $collectionValueType;
-                } elseif ($type->isCollection() && \count($collectionValueType = $type->getCollectionValueTypes()) > 0 && Type::BUILTIN_TYPE_ARRAY === $collectionValueType[0]->getBuiltinType()) {
-                    // get inner type for any nested array
-                    [$innerType] = $collectionValueType;
-
-                    // note that it will break for any other builtinType
-                    $dimensions = '[]';
-                    while (\count($innerType->getCollectionValueTypes()) > 0 && Type::BUILTIN_TYPE_ARRAY === $innerType->getBuiltinType()) {
-                        $dimensions .= '[]';
+                $innerType = $type;
+                if ($type->isCollection() && \count($type->getCollectionValueTypes()) > 0) {
+                    while (1 === \count($innerType->getCollectionValueTypes()) && Type::BUILTIN_TYPE_ARRAY === $innerType->getCollectionValueTypes()[0]->getBuiltinType()) {
                         [$innerType] = $innerType->getCollectionValueTypes();
                     }
 
-                    if (null !== $innerType->getClassName()) {
-                        // the builtinType is the inner one and the class is the class followed by []...[]
-                        $builtinType = $innerType->getBuiltinType();
-                        $class = $innerType->getClassName().$dimensions;
-                    } else {
-                        // default fallback (keep it as array)
-                        $builtinType = $type->getBuiltinType();
-                        $class = $type->getClassName();
+                    $dimensions = '';
+                    $arrayType = $type;
+                    do {
+                        $dimensions .= '[]';
+                        [$arrayType] = $arrayType->getCollectionValueTypes();
+                    } while (\count($arrayType->getCollectionValueTypes()) > 0 && Type::BUILTIN_TYPE_ARRAY === $arrayType->getBuiltinType());
+
+                    if (\count($innerType->getCollectionValueTypes()) > 1 || \in_array($innerType->getCollectionValueTypes()[0]->getBuiltinType(), [Type::BUILTIN_TYPE_OBJECT, Type::BUILTIN_TYPE_ARRAY], true)) {
+                        $builtinType = Type::BUILTIN_TYPE_OBJECT;
+                        $class = $arrayType->getClassName().$dimensions;
+                        $context['value_type'] = $type;
+                        $expectedTypes['array<'.implode('|', array_map(fn (Type $t) => $t->getClassName() ?? $t->getBuiltinType(), $innerType->getCollectionValueTypes())).'>'] = true;
                     }
-                } else {
-                    $builtinType = $type->getBuiltinType();
-                    $class = $type->getClassName();
                 }
 
-                $expectedTypes[Type::BUILTIN_TYPE_OBJECT === $builtinType && $class ? $class : $builtinType] = true;
+                if (!str_ends_with($class, '[]')) {
+                    $expectedTypes[Type::BUILTIN_TYPE_OBJECT === $builtinType && $class ? $class : $builtinType] = true;
+                }
 
                 if (Type::BUILTIN_TYPE_OBJECT === $builtinType) {
                     if (!$this->serializer instanceof DenormalizerInterface) {