+++ src/Symfony/Component/Serializer/Normalizer/ArrayDenormalizer.php
@@ -13,7 +13,9 @@ namespace Symfony\Component\Serializer\Normalizer;
 
 use Symfony\Component\PropertyInfo\Type;
 use Symfony\Component\Serializer\Exception\BadMethodCallException;
+use Symfony\Component\Serializer\Exception\ExtraAttributesException;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
+use Symfony\Component\Serializer\Exception\MissingConstructorArgumentsException;
 use Symfony\Component\Serializer\Exception\NotNormalizableValueException;
 
 /**
@@ -50,17 +52,23 @@ class ArrayDenormalizer implements ContextAwareDenormalizerInterface, Denormaliz
             throw new BadMethodCallException('Please set a denormalizer before calling denormalize()!');
         }
         if (!\is_array($data)) {
-            throw NotNormalizableValueException::createForUnexpectedDataType(sprintf('Data expected to be "%s", "%s" given.', $type, get_debug_type($data)), $data, [Type::BUILTIN_TYPE_ARRAY], $context['deserialization_path'] ?? null);
+            $valueType = $context['value_type'] ?? null;
+            $expected = $valueType ? 'array<'.implode('|', array_map(fn (Type $type) => $type->getClassName() ?? $type->getBuiltinType(), $valueType->getCollectionValueTypes())).'>' : $type;
+
+            throw NotNormalizableValueException::createForUnexpectedDataType(sprintf('Data expected to be "%s", "%s" given.', $expected, get_debug_type($data)), $data, [Type::BUILTIN_TYPE_ARRAY], $context['deserialization_path'] ?? null);
         }
         if (!str_ends_with($type, '[]')) {
             throw new InvalidArgumentException('Unsupported class: '.$type);
         }
 
         $type = substr($type, 0, -2);
+        $valueType = $context['value_type'] ?? null;
 
-        $builtinTypes = array_map(static function (Type $keyType) {
-            return $keyType->getBuiltinType();
-        }, \is_array($keyType = $context['key_type'] ?? []) ? $keyType : [$keyType]);
+        if ($valueType instanceof Type && \count($keyTypes = $valueType->getCollectionKeyTypes()) > 0) {
+            $builtinTypes = array_map(static fn (Type $keyType) => $keyType->getBuiltinType(), $keyTypes);
+        } else {
+            $builtinTypes = array_map(static fn (Type $keyType) => $keyType->getBuiltinType(), \is_array($keyType = $context['key_type'] ?? []) ? $keyType : [$keyType]);
+        }
 
         foreach ($data as $key => $value) {
             $subContext = $context;
@@ -68,7 +76,34 @@ class ArrayDenormalizer implements ContextAwareDenormalizerInterface, Denormaliz
 
             $this->validateKeyType($builtinTypes, $key, $subContext['deserialization_path']);
 
-            $data[$key] = $this->denormalizer->denormalize($value, $type, $format, $subContext);
+            if ($valueType instanceof Type) {
+                foreach ($valueType->getCollectionValueTypes() as $subtype) {
+                    try {
+                        $subContext['value_type'] = $subtype;
+
+                        if ($subtype->isNullable() && null === $value) {
+                            $data[$key] = null;
+
+                            continue 2;
+                        }
+
+                        if (Type::BUILTIN_TYPE_ARRAY === $subtype->getBuiltinType()) {
+                            $class = $type;
+                        } else {
+                            $class = $subtype->getClassName() ?? $subtype->getBuiltinType();
+                        }
+
+                        $data[$key] = $this->denormalizer->denormalize($value, $class, $format, $subContext);
+
+                        continue 2;
+                    } catch (NotNormalizableValueException|InvalidArgumentException|ExtraAttributesException|MissingConstructorArgumentsException $e) {
+                    }
+                }
+
+                throw $e;
+            } else {
+                $data[$key] = $this->denormalizer->denormalize($value, $type, $format, $subContext);
+            }
         }
 
         return $data;
@@ -94,9 +129,6 @@ class ArrayDenormalizer implements ContextAwareDenormalizerInterface, Denormaliz
         return $this->denormalizer instanceof CacheableSupportsMethodInterface && $this->denormalizer->hasCacheableSupportsMethod();
     }
 
-    /**
-     * @param mixed $key
-     */
     private function validateKeyType(array $builtinTypes, $key, string $path): void
     {
         if (!$builtinTypes) {